auth_service/presentation/views/auth_views.py
from django.core.exceptions import ValidationError
from rest_framework import status
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import AllowAny
from rest_framework.response import Response

from application.register_user_service import RegisterUserService
from presentation.serializers.auth_serializers import RegisterUserSerializer, UserSerializer


@api_view(['POST'])
@permission_classes([AllowAny])
def register(request):
    serializer = RegisterUserSerializer(data=request.data)
    if not serializer.is_valid():
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)

    try:
        service = RegisterUserService()
        user = service.register(serializer.validated_data)

        user_serializer = UserSerializer(user)
        return Response({
            'message': 'User registered successfully',
            'user': user_serializer.data
        }, status=status.HTTP_201_CREATED)
    except ValidationError as e:
        return Response({'non_field_errors': [str(e.message)]}, status=status.HTTP_400_BAD_REQUEST)
    except Exception:
        return Response({'error': 'Registration failed'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
